+++ sdk/cards/src/internal/utils/JsonToModelMapper.ts
@@ -81,6 +81,8 @@ import {
     keyWidgets
 } from "../Constants";
 import { assertUnsupportedError, getEnumByName } from "../utils/Util";
+import StaticImageType from "../../model/enums/StaticImageType";
+import { MoEAccessibilityData, KEY_ACCESSIBILITY } from "react-native-moengage";
 
 export function actionFromJson(json: { [k: string]: any }): Action {
     const navigationType: ActionType = getEnumByName(ActionType, json[keyActionType] ?? "");
@@ -171,7 +173,8 @@ export function cardInfoFromJson(json: { [k: string]: any }): CardInfo {
         Array.from((json[keyCards] ?? []) as Iterable<{ [k: string]: any }>)
             .map((cardJson) => {
                 return cardFromJson(cardJson);
-            })
+            }),
+        staticImageFromJson(json[KEY_ACCESSIBILITY] ?? undefined),
     );
 }
 
@@ -181,7 +184,8 @@ export function cardsDataFromJson(json: { [k: string]: any }): CardsData {
         Array.from((json[keyCards] ?? []) as Iterable<{ [k: string]: any }>)
             .map((cardObject) => {
                 return cardFromJson(cardObject);
-            })
+            }),
+        staticImageFromJson(json[KEY_ACCESSIBILITY] ?? undefined),
     );
 }
 
@@ -262,6 +266,17 @@ export function widgetFromJson(json: { [k: string]: any }): Widget {
         Array.from((json[keyActions] ?? []) as Iterable<{ [k: string]: any }>)
             .map((action) => {
                 return actionFromJson(action);
-            })
+            }),
+        json[KEY_ACCESSIBILITY] != null ? MoEAccessibilityData.fromJson(json[KEY_ACCESSIBILITY]) : undefined,
     );
+}
+
+export function staticImageFromJson( json: { [k: string]: any }): { [key in StaticImageType]: MoEAccessibilityData } {
+    const result: { [key in StaticImageType]?: MoEAccessibilityData } = {};
+    for (const type of Object.values(StaticImageType)) {
+        if (json[type]) {
+            result[type as StaticImageType] = MoEAccessibilityData.fromJson(json[type]);
+        }
+    }
+    return result as { [key in StaticImageType]: MoEAccessibilityData };
 }